backend/pyproject.toml
[project]
name = "productflow-backend"
version = "0.1.0"
description = "Backend API for ProductFlow"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16.5",
  "dramatiq[redis]>=1.18.0",
  "fastapi>=0.115.12",
  "httpx>=0.28.1",
  "itsdangerous>=2.2.0",
  "openai>=1.107.1",
  "pillow>=11.2.1",
  "psycopg[binary]>=3.2.9",
  "pydantic-settings>=2.10.1",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.40",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1",
  "pytest-mock>=3.14.1",
  "ruff>=0.12.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/productflow_backend"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]
